Fix error in serializable
